# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontiTOGO API",
        "description": "API RESTful de la plateforme TontiTOGO — Gestion numérique des tontines au Togo avec suivi des cotisations et notifications SMS.",
        "contact": {
            "name": "Support TontiTOGO",
            "email": "support@tontitogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/agent/login": {
            "post": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Connexion de l'agent par numéro de téléphone",
                "operationId": "40155bd0158a0fb398ac2a9539f4ca6b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "22890123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie — retourne user + token"
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    },
                    "403": {
                        "description": "Compte suspendu"
                    }
                }
            }
        },
        "/api/agent/logout": {
            "post": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Déconnexion de l'agent",
                "operationId": "38d81e3a8a1c212393f38b315aa39ec0",
                "responses": {
                    "200": {
                        "description": "Déconnecté"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agent/me": {
            "get": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Profil de l'agent connecté + stats du jour",
                "operationId": "810143d3f358218f0c41f20e24343409",
                "responses": {
                    "200": {
                        "description": "Profil + stats du jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agent/password/change": {
            "post": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Changement de mot de passe obligatoire à la première connexion",
                "operationId": "4e53a415889e0f081adba6f0a88d20bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe changé avec succès"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agent/password/forgot": {
            "post": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Demande de réinitialisation de mot de passe par SMS (OTP)",
                "operationId": "02f1db15f9660f0c14a96cdcd6fda3bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Numéro de téléphone togolais de l'agent",
                                        "type": "string",
                                        "example": "22890123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SMS envoyé (ou numéro inconnu — même réponse par sécurité)"
                    },
                    "422": {
                        "description": "Numéro invalide"
                    }
                }
            }
        },
        "/api/agent/password/reset": {
            "post": {
                "tags": [
                    "Auth Agent"
                ],
                "summary": "Réinitialisation du mot de passe avec le code OTP reçu par SMS",
                "operationId": "0a7cdf4dba4c3e59fd9c84e5197a8786",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "22890123456"
                                    },
                                    "otp": {
                                        "description": "Code à 6 chiffres reçu par SMS",
                                        "type": "string",
                                        "example": "482916"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nouveauMdp123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "nouveauMdp123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé avec succès"
                    },
                    "422": {
                        "description": "Code OTP invalide, expiré ou données incorrectes"
                    },
                    "404": {
                        "description": "Agent introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}